backend/requirements.txt
fastapi==0.96.0
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
pydantic==1.10.8
segment_anything==1.0
torch==2.0.1
torchvision==0.15.2
uvicorn==0.22.0
